python/requirements.txt
face_recognition==1.3.0
Pillow==6.2.2
